General file cleanup

+++ SQLServer2017/tables/Difficulty.sql
@@ -1,14 +1,14 @@
 CREATE TABLE dbo.Difficulty
 (
-	DifficultyID						INT IDENTITY (1, 1) NOT NULL
-	,DifficultyName						NVARCHAR(100) NOT NULL
+   DifficultyID                     TINYINT IDENTITY (1, 1) NOT NULL
+   ,DifficultyName                  NVARCHAR(100) NOT NULL
 );
 
 ALTER TABLE
-	dbo.Difficulty
+   dbo.Difficulty
 ADD CONSTRAINT
-	PK_Difficulty
+   PK_Difficulty
 PRIMARY KEY
 (
-	DifficultyID
+   DifficultyID
 );
